Remove anti-pattern script example
