--- examples/workflows/dags/dag-diamond-with-callable-decorators.yaml
@@ -1,54 +1,0 @@
-apiVersion: argoproj.io/v1alpha1
-kind: Workflow
-metadata:
-  generateName: dag-diamond-
-spec:
-  entrypoint: diamond
-  templates:
-  - dag:
-      tasks:
-      - arguments:
-          parameters:
-          - name: message
-            value: A
-        name: A
-        template: echo
-      - arguments:
-          parameters:
-          - name: message
-            value: B
-        depends: A
-        name: B
-        template: echo
-      - arguments:
-          parameters:
-          - name: message
-            value: C
-        depends: A
-        name: C
-        template: echo
-      - arguments:
-          parameters:
-          - name: message
-            value: D
-        depends: B && C
-        name: D
-        template: echo
-    name: diamond
-  - inputs:
-      parameters:
-      - name: message
-    name: echo
-    script:
-      command:
-      - python
-      image: python:3.12
-      source: |-
-        import os
-        import sys
-        sys.path.append(os.getcwd())
-        import json
-        try: message = json.loads(r'''{{inputs.parameters.message}}''')
-        except: message = r'''{{inputs.parameters.message}}'''
-
-        print(message)
