--- examples/workflows/dags/dag_diamond_with_callable_decorators.py
@@ -1,15 +1,0 @@
-from hera.workflows import DAG, Workflow, script
-
-
-@script(image="python:3.12")
-def echo(message):
-    print(message)
-
-
-with Workflow(generate_name="dag-diamond-", entrypoint="diamond") as w:
-    with DAG(name="diamond"):
-        A = echo(name="A", arguments={"message": "A"})
-        B = echo(name="B", arguments={"message": "B"})
-        C = echo(name="C", arguments={"message": "C"})
-        D = echo(name="D", arguments={"message": "D"})
-        A >> [B, C] >> D
