+++ examples/workflows/misc/suspend-input-duration.yaml
@@ -17,6 +17,13 @@ spec:
         valueFrom:
           supplied: {}
     suspend: {}
+  - inputs:
+      parameters:
+      - default: '10'
+        name: duration
+    name: input-duration-suspend
+    suspend:
+      duration: '{{inputs.parameters.duration}}'
   - name: suspend
     steps:
     - - name: get-value-step
@@ -27,10 +34,3 @@ spec:
             value: '{{steps.get-value-step.outputs.parameters.duration}}'
         name: custom-delay-step
         template: input-duration-suspend
-  - inputs:
-      parameters:
-      - default: '10'
-        name: duration
-    name: input-duration-suspend
-    suspend:
-      duration: '{{inputs.parameters.duration}}'
